+++ lib/src/extensions/extensions_import.dart
@@ -15,5 +15,6 @@ part '../extensions/data/theme_extension.dart';
 part 'data/ui/align_extension.dart';
 part 'data/ui/flex_extension.dart';
 part 'data/ui/padding_extension.dart';
+part 'data/ui/positioned_extension.dart';
 part 'data/ui/sized_box_extension.dart';
 part 'data/ui/visibility_extension.dart';