+++ lib/src/reusable/structure/lazy_indexed_stack.dart
@@ -1,142 +1,149 @@
-import 'package:flutter/widgets.dart';
+part of '../reusable_import.dart';
 
-/// A performance-optimized version of [IndexedStack] that only builds the widget associated with the current index or preloaded indices.
+/// A widget that displays a stack of pages, with lazy loading, fade transitions, and preloading support.
 ///
-/// [LazyIndexedStack] helps improve app performance by deferring the creation of child widgets that are not currently visible.
-/// It ensures that only the visible widget (at [index]) and preloaded widgets are built initially, while the other widgets are unloaded
-/// and replaced with a placeholder [unloadWidget] until they are needed again.
+/// This widget provides an efficient way to display a stack of pages where only the currently
+/// selected page is built, and optionally preloaded pages are kept in the widget tree.
+/// Pages that are not visible are replaced with a placeholder widget to reduce unnecessary rebuilds.
 ///
-/// This widget is especially useful in scenarios where the stack has a large number of children, but only a few are visible at any
-/// given time, making it ideal for scenarios like tab navigation, where only one widget is visible at a time.
+/// It also includes a fade transition between page switches, adding a smooth and professional animation effect.
 ///
-/// Example:
-/// ```dart
-/// LazyIndexedStack(
-///  index: _currentIndex,
-///  preloadIndexes: [0, 1, 2],
-///  children: [
-///  Tab1(),
-///  Tab2(),
-///  Tab3(),
-///  Tab4(),
-///  Tab5(),
-///  ],
-///  unloadWidget: Container(),
-///  )
-///  ```
+/// ### Key Features:
+/// - **Lazy Loading**: Only the selected page and preloaded pages are built. Non-visible pages are replaced with a placeholder widget (`unloadWidget`).
+/// - **Fade Transition**: A fade effect is applied when switching between pages, making the transition smooth.
+/// - **Preloading Support**: You can specify a list of indices (`preloadIndexes`) to preload pages that are likely to be needed soon.
+/// - **Efficient Rebuilding**: Pages that are not in the current index or preloaded are not rebuilt, saving resources and improving performance.
+///
+/// ### Parameters:
+/// - [index]: The current index of the page to be displayed. If not provided, defaults to `0`.
+/// - [length]: The total number of pages.
+/// - [duration]: The duration of the fade transition animation (defaults to 800ms).
+/// - [pageBuilder]: A function that takes an index and returns the widget to be displayed at that index.
+/// - [unloadWidget]: A widget that replaces non-visible pages in the widget tree (default is `SizedBox.shrink()`).
 class LazyIndexedStack extends StatefulWidget {
-  /// The widget to display in place of unloaded children.
-  ///
-  /// The [unloadWidget] is shown when a child widget is not visible. By default, it is a [Container].
+  /// The index of the currently visible page. Defaults to `0`.
+  final int? index;
+
+  /// The total number of pages.
+  final int? length;
+
+  /// The duration of the fade transition animation. Defaults to 800 milliseconds.
+  final Duration? duration;
+
+  /// A function that returns the widget for a given index.
+  final Widget? Function(int pageIndex)? pageBuilder;
+
+  /// The widget that replaces unloaded pages. Defaults to an empty container (`SizedBox.shrink()`).
   final Widget unloadWidget;
 
-  /// A list of indices corresponding to the children that should be preloaded.
+  /// Creates a new [LazyIndexedStack] widget.
   ///
-  /// Children at the indices specified here will be built and remain in the widget tree even if they are not visible.
-  /// This can be useful for optimizing performance by ensuring that certain children are always available when their index
-  /// is selected.
-  final List<int> preloadIndexes;
-
-  /// Alignment of the children within the [IndexedStack].
+  /// - The [index] and [length] parameters are required, while the rest are optional.
+  /// - The [pageBuilder] function is required to return a widget for each page index.
+  /// - The [unloadWidget] parameter is optional and defaults to an empty container.
+  /// - The [duration] parameter is optional and defaults to 800 milliseconds.
   ///
-  /// This property determines how the children are aligned within the stack. The default value is [AlignmentDirectional.topStart].
-  final AlignmentGeometry alignment;
-
-  /// The sizing behavior of the stack's children.
-  ///
-  /// Similar to the [StackFit] property of the original [IndexedStack], this controls how the children inside the stack should
-  /// size themselves relative to each other.
-  final StackFit sizing;
-
-  /// The text direction in which the children of the stack should be laid out.
-  ///
-  /// If `null`, the text direction will be inherited from the ambient [Directionality].
-  final TextDirection? textDirection;
-
-  /// The index of the currently visible child in the stack.
-  ///
-  /// Only the child at this index will be built immediately. Other children will either be unloaded or replaced with the
-  /// [unloadWidget] unless they are part of the preloaded indices.
-  final int index;
-
-  /// The children widgets of the stack.
-  ///
-  /// These widgets will be displayed in a stack-like manner, but only the widget corresponding to the [index] or preloaded
-  /// indices will be built immediately. Other widgets will be replaced with the [unloadWidget] until needed.
-  final List<Widget> children;
-
-  /// Creates a [LazyIndexedStack] that wraps a [IndexedStack], deferring the building of non-visible widgets.
-  ///
-  /// - The widget at the [index] will be built immediately.
-  /// - The widgets in the [preloadIndexes] list will also be preloaded and available when needed.
-  /// - All other widgets are replaced with [unloadWidget], which defaults to [Container].
-  LazyIndexedStack({
+  /// ### Example Usage:
+  /// ```dart
+  /// LazyIndexedStack(
+  ///   index: selectedIndex,
+  ///   length: 5,
+  ///   duration: const Duration(milliseconds: 400),
+  ///   preloadIndexes: [0, 2], // Preload pages 0 and 2
+  ///   unloadWidget: const SizedBox.shrink(), // Placeholder for unloaded pages
+  ///   pageBuilder: (index) {
+  ///     switch (index) {
+  ///       case 0:
+  ///         return Text('Page 1').center;
+  ///       case 1:
+  ///         return Text('Page 2').center;
+  ///       case 2:
+  ///         return HomeScreen(bloc: HomeBloc());
+  ///       case 3:
+  ///         return Text('Page 4').center;
+  ///       case 4:
+  ///         return Text('Page 5').center;
+  ///       default:
+  ///         return const SizedBox.shrink(); // Fallback
+  ///     }
+  ///   },
+  /// )
+  /// ```
+  /// In this example:
+  /// - The `selectedIndex` will determine which page is shown.
+  /// - Non-visible pages will be replaced with an empty `SizedBox`.
+  /// - The `pageBuilder` is used to define custom widgets for each page.
+  const LazyIndexedStack({
     super.key,
-    Widget? unloadWidget,
-    this.preloadIndexes = const [],
-    this.alignment = AlignmentDirectional.topStart,
-    this.sizing = StackFit.loose,
-    this.textDirection,
-    required this.index,
-    required this.children,
-  }) : unloadWidget = unloadWidget ?? Container();
+    this.index,
+    this.length,
+    this.duration = const Duration(milliseconds: 400),
+    this.pageBuilder,
+    this.unloadWidget = const SizedBox.shrink(),
+  });
 
   @override
-  LazyIndexedStackState createState() => LazyIndexedStackState();
+  _LazyIndexedStackState createState() => _LazyIndexedStackState();
 }
 
-class LazyIndexedStackState extends State<LazyIndexedStack> {
-  late List<Widget> _children;
-  final _stackKey = GlobalKey();
+class _LazyIndexedStackState extends State<LazyIndexedStack>
+    with SingleTickerProviderStateMixin {
+  AnimationController? _controller;
+  PageController? _pageController;
+
+  @override
+  void didUpdateWidget(LazyIndexedStack oldWidget) {
+    super.didUpdateWidget(oldWidget);
+
+    // Animate and change page only if the index changes
+    if (widget.index != oldWidget.index) {
+      _controller!.forward(from: 0.0);
+      _pageController!.animateToPage(
+        widget.index!,
+        duration: const Duration(milliseconds: 1),
+        curve: Curves.easeInToLinear,
+      );
+    }
+  }
 
   @override
   void initState() {
     super.initState();
-    _children = _initializeChildren();
+    // Initialize the animation controller and page controller
+    _controller = AnimationController(
+      vsync: this,
+      duration: widget.duration,
+    )..forward();
+    // Initialize the page controller
+    _pageController = PageController(initialPage: widget.index ?? 0);
   }
 
   @override
-  void didUpdateWidget(final LazyIndexedStack oldWidget) {
-    super.didUpdateWidget(oldWidget);
-
-    // Reinitialize children if the number of children changes.
-    if (widget.children.length != oldWidget.children.length) {
-      _children = _initializeChildren();
-    }
-
-    // Update the widget at the current index if it has changed.
-    _children[widget.index] = widget.children[widget.index];
+  void dispose() {
+    // Dispose the controllers
+    _controller!.dispose();
+    _pageController!.dispose();
+    super.dispose();
   }
 
   @override
-  Widget build(final BuildContext context) {
-    return IndexedStack(
-      key: _stackKey,
-      index: widget.index,
-      alignment: widget.alignment,
-      textDirection: widget.textDirection,
-      sizing: widget.sizing,
-      children: _children,
+  Widget build(BuildContext context) {
+    return FadeTransition(
+      opacity: _controller!,
+      child: PageView.builder(
+        controller: _pageController,
+        itemCount: widget.length,
+        physics: const NeverScrollableScrollPhysics(),
+        itemBuilder: (BuildContext context, int itemIndex) {
+          // Build only the selected page and preloaded pages
+          if (itemIndex == widget.index) {
+            return widget.pageBuilder!(itemIndex)!;
+          } else {
+            return widget
+                .unloadWidget; // Replace unloaded pages with placeholder
+          }
+        },
+      ),
     );
-  }
-
-  /// Initializes the list of children by replacing widgets that are not visible with [unloadWidget].
-  ///
-  /// This method ensures that:
-  /// - The child at the [index] is built immediately.
-  /// - The children in [preloadIndexes] are built and remain in the widget tree.
-  /// - Other children are replaced with [unloadWidget] until they are needed again.
-  List<Widget> _initializeChildren() {
-    return widget.children.asMap().entries.map((entry) {
-      final index = entry.key;
-      final childWidget = entry.value;
-
-      // Only build the widget if its index is the current [index] or part of [preloadIndexes].
-      if (index == widget.index || widget.preloadIndexes.contains(index)) {
-        return childWidget;
-      } else {
-        return widget.unloadWidget;
-      }
-    }).toList();
   }
 }